Rename filters folder to config
