+++ src/config/game.ts
