+++ src/config/inputs.ts
