+++ src/config/stocks.ts
